requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.11.0
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.2.post1
arxiv==2.1.3
attrs==24.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
blinker==1.9.0
Brotli==1.1.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==43.0.3
cssselect2==0.7.0
dataclasses-json==0.6.7
distro==1.9.0
docopt==0.6.2
emoji==2.14.0
eval_type_backport==0.2.0
fastapi==0.115.5
feedparser==6.0.11
filetype==1.2.0
fonttools==4.54.1
frozenlist==1.5.0
gitdb==4.0.11
GitPython==3.1.43
gpt-researcher==0.10.3
greenlet==3.1.1
h11==0.14.0
html5lib==1.1
htmldocx==0.0.6
httpcore==1.0.6
httpx==0.27.2
httpx-sse==0.4.0
idna==3.10
Jinja2==3.1.4
jiter==0.7.1
joblib==1.4.2
json5==0.9.28
json_repair==0.30.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain==0.3.7
langchain-community==0.3.7
langchain-core==0.3.18
langchain-ollama==0.2.0
langchain-openai==0.2.8
langchain-text-splitters==0.3.2
langdetect==1.0.9
langsmith==0.1.143
loguru==0.7.2
lxml==5.3.0
lxml_html_clean==0.4.0
Markdown==3.7
markdown-it-py==3.0.0
markdown2==2.5.1
MarkupSafe==3.0.2
marshmallow==3.23.1
md2pdf==1.0.1
mdurl==0.1.2
mistune==3.0.2
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.13.5
nest-asyncio==1.6.0
nltk==3.9.1
numpy==1.26.4
olefile==0.47
ollama==0.3.3
openai==1.54.4
orjson==3.10.11
packaging==24.2
pandas==2.2.3
pillow==10.4.0
propcache==0.2.0
protobuf==5.28.3
psutil==6.1.0
pyarrow==18.0.0
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
pydeck==0.9.1
pydyf==0.11.0
Pygments==2.18.0
PyMuPDF==1.24.13
pypdf==5.1.0
pyphen==0.15.0
python-dateutil==2.8.2
python-docx==1.1.2
python-dotenv==1.0.1
python-iso639==2024.10.22
python-magic==0.4.27
python-multipart==0.0.17
python-oxmsg==0.0.1
pytz==2024.2
PyYAML==6.0.2
RapidFuzz==3.10.1
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.21.0
sgmllib3k==1.0.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.35
starlette==0.41.2
streamlit==1.40.1
tenacity==9.0.0
tiktoken==0.8.0
tinycss2==1.4.0
tinyhtml5==2.0.0
toml==0.10.2
tornado==6.4.1
tqdm==4.67.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
unstructured==0.16.5
unstructured-client==0.27.0
urllib3==2.2.3
watchdog==6.0.0
weasyprint==63.0
webencodings==0.5.1
websockets==14.1
win32-setctime==1.1.0
wrapt==1.16.0
yarl==1.17.1
zopfli==0.2.3.post1
